+++ src/main/java/systemData/services/SystemDataServices/IncidentTypeService.java
@@ -1,6 +1,9 @@
 package systemData.services.SystemDataServices;
 
+import javassist.NotFoundException;
+import systemData.models.BasicData.Incident.Incident;
 import systemData.models.BasicData.IncidentType.IncidentType;
+import systemData.repos.SystemDataRepos.IncidentRepo;
 import systemData.repos.SystemDataRepos.IncidentTypeRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,7 +15,8 @@ import java.util.Optional;
 public class IncidentTypeService {
     @Autowired
     IncidentTypeRepo incidentTypeRepo;
-
+    @Autowired
+    IncidentRepo incidentRepo;
     public List<IncidentType> getAllTypes(){
         List<IncidentType> incidentTypes = incidentTypeRepo.findAll();
         if(incidentTypes.isEmpty()) return null;
@@ -39,9 +43,11 @@ public class IncidentTypeService {
 
     }
 
-    public boolean DeleteOne(String code){
+    public boolean DeleteOne(String code) throws NotFoundException {
         Optional<IncidentType> checked =  incidentTypeRepo.findById(code);
         if(!checked.isPresent()) return false;
+        List<Incident> incidents = incidentRepo.findByIncidentTypeCode(checked.get());
+        if(!incidents.isEmpty()) throw new NotFoundException("Can't delete this type, there are incidents with this type");
         incidentTypeRepo.deleteById(code);
         return true;
